requirements.txt
black
numpy
pytest
coverage
pandas
plotly
kaleido
scikit-learn
biopython
requests